ENH: create vector.py and matrix.py files
